refactor: refactoring about effects
